feat: add application config
- move env functions to separate file
- add Config structure
- add method to read config using consul client
- add function for consul watch endpoint
